Reservation 오류수정중

+++ src/main/java/com/dw/library/mapper/ReservationMapper.java
@@ -15,6 +15,6 @@ public interface ReservationMapper {
 
     Reservation findByReservationId(@Param("reservationId") Long reservationId);
 
-    int deleteReservation(@Param("reservationId") Long reservationId);
+    long deleteReservation(@Param("reservationId") Long reservationId);
 
 }
